Value copy functions -> value.c
